+++ app/lib/conferences/eventCall.js
@@ -1,8 +1,7 @@
 import axios from "axios";
 
 const eventUrl = process.env.NEXT_PUBLIC_EVENT_BACKEND_URL;
-const nextDeployUrl =
-  process.env.NEXT_PUBLIC_API_URL || "http://localhost:3000";
+const nextDeployUrl = "http://localhost:3000";
 
 //NextJS local API route call begins
 export const signCook = async (mail) => {
@@ -148,12 +147,12 @@ export const getAllEvents = async (eid) => {
   return res;
 };
 
-export const getUserEventDeatils = async (uid, auth) => {
+export const getUserEventDeatilsByState = async (uid, auth,state) => {
   const headers = {
     Accept: "application/vnd.api+json",
     Authorization: `JWT ${auth}`,
   };
-  const res = await axios.get(`${eventUrl}/v1/users/${uid}/events`, {
+  const res = await axios.get(`${eventUrl}/v1/users/${uid}/events?filter=[{"name" : "state","op" : "eq","val" : "${state}"}]`, {
     headers: headers,
   });
   return res;
