+++ cms/config/initialData/sub-menus.json
@@ -150,7 +150,7 @@
   {
     "id": 19,
     "label": "Dashboard",
-    "url": "/conferences/admin/dashboard",
+    "url": "/conferences/admin/dashboard/published",
     "published_at": "2021-06-21T09:03:49.991Z",
     "created_at": "2021-06-21T09:03:57.289Z",
     "updated_at": "2021-06-21T09:03:57.289Z"
